adjust controller to accept all rules in crud

+++ app/Http/Controllers/ClientController.php
@@ -5,12 +5,15 @@ namespace App\Http\Controllers;
 use App\Repositories\ClientRepository;
 use Illuminate\Http\Request;
 
-class ClientController extends Controller {
-    public function index() {
-        return ClientRepository::index();
+class ClientController extends Controller
+{
+    public function index()
+    {
+        return response()->json(ClientRepository::index(), 200);
     }
 
-    public function store(Request $request) {
+    public function store(Request $request)
+    {
         $requestBody = $request->json()->all();
 
         $this->validate($request, [
@@ -26,29 +29,66 @@ class ClientController extends Controller {
 
         $clientExists = ClientRepository::getClientByDocument($requestBody['document']);
 
-        if($clientExists) {
+        if ($clientExists) {
             return response()->json([
                 'message' => 'Client already exists'
             ], 409);
         }
 
-        return ClientRepository::store($requestBody);
+        ClientRepository::store($requestBody);
 
+        return response()->json([
+            'message' => 'Client created successfully'
+        ], 201);
     }
 
-    public function show($id) {
-        return ClientRepository::show($id);
+    public function show($id)
+    {
+        return response()->json(ClientRepository::show($id), 200);
     }
 
-    public function update(Request $request, $id) {
+    public function update(Request $request, $id)
+    {
         $requestBody = $request->json()->all();
 
-        return ClientRepository::update($requestBody, $id);
+        $clientExists = ClientRepository::getClientById($id);
+
+        if (!$clientExists) {
+            return response()->json([
+                'message' => 'Client not exists'
+            ], 409);
+        }
+
+        $clientDocumentExists = ClientRepository::getClientByDocument($requestBody['document']);
+
+        if ($clientDocumentExists) {
+            return response()->json([
+                'message' => 'Client document already exists'
+            ], 409);
+        }
+
+        ClientRepository::update($requestBody, $id);
+
+        return response()->json([
+            'message' => 'Client updated successfully'
+        ], 200);
     }
 
-    public function destroy($id) {
-        return ClientRepository::destroy($id);
-    }
+    public function destroy($id)
+    {
 
-    
+        $clientExists = ClientRepository::getClientById($id);
+
+        if (!$clientExists) {
+            return response()->json([
+                'message' => 'Client not exists'
+            ], 409);
+        }
+
+        ClientRepository::destroy($id);
+
+        return response()->json([
+            'message' => 'Client deleted successfully'
+        ], 200);
+    }
 }
